Refine main screen layout to follow design
